Refactor Profession model
